pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odoodev"
version = "0.1.0"
description = "CLI for Odoo development environment management"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "questionary>=2.0.0",
    "textual>=0.40.0",
]

[project.scripts]
odoodev = "cli.main:app"

[tool.setuptools.packages.find]
include = ["cli*"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["addons"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501", "B008", "B904", "SIM108"]

[tool.ruff.lint.isort]
known-third-party = ["odoo", "odoo.addons"]

[tool.ruff.format]
quote-style = "double"

[tool.pylint.master]
load-plugins = "pylint_odoo"

[tool.pylint."messages control"]
disable = ["C0114", "C0115", "C0116", "W0104"]

[tool.pylint.odoolint]
manifest_required_authors = true
manifest_deprecated_keys = ["description", "active"]
